Added a fat finger box for delete button

+++ src/project3/BankGUI.java
@@ -513,7 +513,10 @@ public class BankGUI extends JFrame {
 			if (event.getSource() == modifyButton[1]) {
 				// delete selected row
 				try {
-					bankModel.delete(table.getSelectedRow());
+					int result = JOptionPane.showConfirmDialog((Component) null, "Are you sure?",
+				        "Confirm", JOptionPane.OK_CANCEL_OPTION);
+					if(result == 0)
+						bankModel.delete(table.getSelectedRow());
 				} catch (Exception e) {
 					// display error message to user
 					JOptionPane.showMessageDialog(null,
